requirements.txt
python3==3.13.3
pandas==2.2.3
python-docx==16
openpyxl==3.1
python_dotenv==1.1.0
requests==2.32.3
flask==3.1.0
PIL
pillow==11.2.11
pip==25.0.1
requests==2.32.3
typing-extensions==4.7.1
numpy==2.2.5
python_dotenv==1.1.0
dateutil==5.1.1
python-docx==16
urllib3==1.26.16
typing-extensions==4.7.1